rename embeddings/luxical/embed_source.py -> pipeline.py
The "embed_source" name was redundant given the embeddings/luxical/
package name and conflicted with the embed_source() function it exports.
Importers now read ...luxical.pipeline import embed_source.
