+++ experiments/datakit/cluster/exp_smoke.py
@@ -41,7 +41,7 @@ from experiments.datakit.cluster.assign import assign_source  # noqa: E402
 from experiments.datakit.cluster.sample import sample_centroid_inputs  # noqa: E402
 from experiments.datakit.cluster.summarize import summarize_at_k  # noqa: E402
 from experiments.datakit.cluster.train import train_centroids  # noqa: E402
-from experiments.datakit.embeddings.luxical.embed_source import (  # noqa: E402
+from experiments.datakit.embeddings.luxical.pipeline import (  # noqa: E402
     LUXICAL_REPO,
     LUXICAL_WEIGHTS_FILE,
     embed_source,
